+++ bin/chr_list_generator.py
