+++ bin/generate_xml.py
