+++ bin/transfer_integrity_check.py
@@ -1,0 +1,281 @@
+#!/usr/bin/env python3
+
+import subprocess
+import pandas as pd
+import argparse
+from argparse import RawTextHelpFormatter
+import os
+import fnmatch
+import glob
+import hashlib
+import gzip
+import sys
+from datetime import datetime
+
+#run on codon
+###include classes, functions, error handling (incl. sys.exits), raise exceptions where necessary
+
+#TODO: fix paths in script
+
+# -------- #
+# Parsing arguments
+# -------- #
+description = """
+This script transfers all files from a user's COVID upload area to the drag and drop directory on codon (or a user specified directory)
+and compares checksums to verify file integrity. 
+"""
+
+example = """
+Example 1:
+    python3 transfer_integrity_check.py -u 65352cc8-c581-43e0-8451-8eddbfd39a5d 
+    
+Example 2:
+    python3 transfer_integrity_check.py -u 65352cc8-c581-43e0-8451-8eddbfd39a5d -o <my output dir>
+    
+"""
+
+parser = argparse.ArgumentParser(description=description, epilog=example, formatter_class=RawTextHelpFormatter)
+
+parser.add_argument('-u', '--UUID', help='User specific access key for drag and drop tool', type=str, required=True)
+
+# adding optional output argument for future use/flexibility. Default output directory = /nfs/production/cochrane/ena/users/covid/covid/{args.UUID}'
+parser.add_argument('-o', '--output', help='Specify output directory to transfer files from S3 bucket to, and hold files + subdirs generated by the script', type=str, required=False) #False #No trailing slash
+
+
+args = parser.parse_args()
+
+def run_bash_cmds(fname):
+    """ Function to run specific bash commands for upload, transfer, or to calculate md5
+    """
+    '''
+    if cmds == upload_cmds:
+        with open(f'{outdir}/{trans_int_outdir}/upload_area_stdout.txt', 'w+') as stdout:
+            subprocess.run(upload_cmds, shell=True, stdin=subprocess.PIPE, stdout=stdout, stderr=subprocess.PIPE, encoding='utf8') #universal_newlines to accept input/output as str not bytes
+            stdout.seek(0)
+            print(stdout.read())
+
+    elif cmds == transfer_cmds:
+        with open(f'{outdir}/{trans_int_outdir}/file_transfer_log.txt', 'w+') as logfile:
+            subprocess.run(transfer_cmds, shell=True, stdin=subprocess.PIPE, stdout=logfile, stderr=logfile, encoding='utf8')
+            logfile.seek(0)
+            print(logfile.read())
+
+    elif cmds == md5sum:
+    '''
+    if args.output:
+        outdir = f"{args.output}"  # user specified outdir
+
+    else:
+        outdir = f"/nfs/production/cochrane/ena/users/covid/covid/{args.UUID}"
+    with open(f'{outdir}/{fname}.md5', 'w+') as fmd5:
+        subprocess.run(f"md5sum '{outdir}/{fname}'", stdout=fmd5, shell=True, stderr=subprocess.PIPE, stdin=subprocess.PIPE, encoding='utf8') # adding '' to account for file name
+        fmd5.seek(0)  # to move cursor back to start of file for reading
+        run_bash_cmds.fdata = fmd5.read()
+        print(run_bash_cmds.fdata)  # to view contents of each .md5 file
+    '''
+    else:
+        raise ValueError("ERROR with input bash commands, exiting", file=sys.stderr)
+        sys.exit()
+    '''
+
+
+
+def transfer():
+    # -------- #
+    # Hard-coded variables
+    # -------- #
+    # codon_dir = '/nfs/production/cochrane/ena/users/covid/covid' #'/nfs/production/cochrane/ena/users/covid/covid'
+    # path = f"{codon_dir}/{args.UUID}/"
+    # Data file extensions accepted by tool
+    file_exts = ('*fasta*','*fastq*','*cram*','*bam*')
+    sheet_exts = ('*.xls*','*.tsv*','*.csv*','*.txt*')
+    # Name of the columns in the summary dataframe
+    file_name = "file_name"
+    uploaded_md5_val = "uploaded_md5_value"
+    new_md5_val = "new_md5_value"
+    integrity_check = "file_integrity_check"
+    # Datetime string for summary tsv file
+    now = datetime.now()
+    dt_string = now.strftime("%y-%m-%d_%H-%M-%S")
+
+
+    # -------- #
+    # Code blocks
+    # -------- #
+
+
+    if args.output:
+        outdir = f"{args.output}"  # user specified outdir
+
+    else:
+        outdir = f"/nfs/production/cochrane/ena/users/covid/covid/{args.UUID}"
+
+    trans_int_outdir = "transfer_integrity_output" #subdir for output files
+
+    # Create outdir for UUID transfer + trans_int_outdir for output files, and/or check if it already exists
+
+    dir_exists = os.path.isdir(outdir)  # T or F
+    #print(dir_exists)
+    if not dir_exists and outdir != '':
+        try:
+            os.makedirs(f"{outdir}/{trans_int_outdir}")
+            print(f"Successfully created output directory {outdir}/{trans_int_outdir}")
+        except Exception as error:
+            print(error)
+            print(f"\n ERROR in transfer_integrity_check.py: could not create output directory '{outdir}/{trans_int_outdir}'. Exiting.", file=sys.stderr)
+            sys.exit()
+    if not os.path.exists(f"{outdir}/{trans_int_outdir}"):
+        os.makedirs(f"{outdir}/{trans_int_outdir}")
+
+
+
+    # Select and list files associated to specified UUID
+    upload_cmds = f"covid-util select {args.UUID}; covid-util list"
+    #run_bash_cmds(upload_cmds)
+    with open(f'{outdir}/{trans_int_outdir}/upload_area_stdout.txt', 'w+') as stdout:
+        subprocess.run(upload_cmds, shell=True, stdin=subprocess.PIPE, stdout=stdout, stderr=subprocess.PIPE,
+                       encoding='utf8')  # universal_newlines to accept input/output as str not bytes
+        stdout.seek(0)
+        print(stdout.read())
+
+    # Transfer files from S3 bucket to output dir
+    if args.output:
+        transfer_cmds = f"cd {outdir} && aws --endpoint-url https://uk1s3.embassy.ebi.ac.uk s3 cp s3://covid-utils-ui-88560523/{args.UUID} . --recursive"
+    else:
+        transfer_cmds = f"cd {os.path.dirname(outdir)}; cd venv/; source bin/activate && cd {outdir} && aws --endpoint-url https://uk1s3.embassy.ebi.ac.uk s3 cp s3://covid-utils-ui-88560523/{args.UUID} . --recursive" # default dir on codon
+
+    # do not need to run aws configure each time to interact w/ aws cli API
+    #run_bash_cmds(transfer_cmds)
+    with open(f'{outdir}/{trans_int_outdir}/file_transfer_log.txt', 'w+') as logfile:
+        subprocess.run(transfer_cmds, shell=True, stdin=subprocess.PIPE, stdout=logfile, stderr=logfile,
+                       encoding='utf8')
+        logfile.seek(0)
+        print(logfile.read())
+
+    # # Check all files have transferred correctly
+    with open(f'{outdir}/{trans_int_outdir}/upload_area_stdout.txt', 'r') as stdout:
+        total_upload = (stdout.readlines()[-1]) #prints last line of stdout = no. of items in upload area #14 items
+    print("files uploaded =", total_upload)
+
+    total_transfer = len([name for name in os.listdir(outdir) if os.path.isfile(os.path.join(outdir, name))]) #list no. of files in specified dir
+    total_transfer = str(total_transfer) + " items"
+    print(f"files transferred to {outdir} =", total_transfer) #TODO: remember if the script is re-run, this statement will not be accurate as passed files will be deleted
+
+    # # # -------- #
+    # # # File integrity check
+    # # # -------- #
+
+    #os.chdir(outdir) #changing wd to output dir (or if not specified, default dir)
+
+    submitted_files = []
+
+    # Creating summary df to record integrity checks
+    summary_df = pd.DataFrame(columns = [file_name, uploaded_md5_val, new_md5_val, integrity_check])
+
+    # Check if PASS and FAIL dirs have already been created (i.e if this is a re-run of script or not)
+    # If yes, remove pass files from UUID dir, if not, create the empty subdirs
+    pdir, fdir = os.path.join(f'{outdir}/{trans_int_outdir}', "pass"), os.path.join(f'{outdir}/{trans_int_outdir}', "fail")
+    # pdir, fdir = f'{outdir}/{trans_int_outdir}/pass', f'{outdir}/{trans_int_outdir}/fail'
+    pdir_exists, fdir_exists = os.path.isdir(pdir), os.path.isdir(fdir) # T or F
+
+    if pdir_exists and fdir_exists:
+        for file in [f for f in os.listdir(outdir) if os.path.isfile(f)]: #removes dirs from os.lsitdir(path) output
+            if any(fnmatch.fnmatch(file, f'*{pfile}*') for pfile in os.listdir(pdir)):
+                os.remove(f'{file}')  # removing pass files from UUID dir
+    elif not pdir_exists and not fdir_exists:
+        os.makedirs(pdir), os.makedirs(fdir)
+    else:
+        raise AssertionError("ERROR creating PASS and FAIL subdirectories, see logfile", file=sys.stderr)
+
+    # Creating directory for drag and drop spreadsheet & moving latest version here
+    if not os.path.exists(f'{outdir}/{trans_int_outdir}/input_spreadsheet'):
+        os.mkdir(f'{outdir}/{trans_int_outdir}/input_spreadsheet') #creates single child dir over mkdirs
+
+    i_sheets = []
+    for sh_ext in sheet_exts:
+        i_sheets += glob.glob(f'{outdir}/{sh_ext}')
+    print(i_sheets)
+
+    datestrings = sorted([sheet.split(".")[-2] for sheet in i_sheets])  # extracts file datestrings (e.g. 2023-01-25T12:45:08) into sorted list
+    latest_ds = str(datestrings[-1])  # extracts most recent datestring
+    latest_file = (str([sheet for sheet in i_sheets if latest_ds in sheet])[1:-1]).strip('\'') # removing [] and '' from list element
+
+    try:
+        os.rename(f'{latest_file}', f'{outdir}/{trans_int_outdir}/input_spreadsheet/{os.path.basename(latest_file)}')
+
+    except Exception as error:
+        print(f"ERROR in transferring file to input_spreadsheet directory. Exiting", file=sys.stderr)
+        print(error)
+        sys.exit()
+
+    # Obtaining a list of data files to submit
+    for fname in os.listdir(outdir):
+        if any(fnmatch.fnmatch(fname, ext) for ext in file_exts):
+            print(fname)
+            submitted_files.append(fname)
+
+            # calculating md5 for each data file, and storing each in corresponding .md5 file
+            run_bash_cmds(fname)
+
+            # # isolating original and newly calculated hashes
+            new_md5 = run_bash_cmds.fdata.split()[0] # extracts first col from .md5 file
+            original_md5 = (os.path.splitext(f"{fname}")[1])[1:] # extracts '.<hash>' from uploaded filename, and removes the leading '.'
+            print(f"new md5 = ", new_md5), print(f"original md5 = ", original_md5, end='\n\n') #print blank line at the end of print statement
+
+            # rename submitted files to remove the hash (for downstream Webin-CLI submission)
+            new_fname = os.path.splitext(f"{fname}")[0]
+            #print(new_fname)
+            # os.rename(f"{outdir}/{fname}", new_fname)
+            os.rename(f"{outdir}/{fname}", f"{outdir}/{new_fname}") # to overcome OSError: [Errno 18] Invalid cross-device link error
+
+            # We add the output of the pushed changes to the summary dataframe
+            new_df = pd.DataFrame([{file_name: new_fname, uploaded_md5_val: original_md5, new_md5_val: new_md5, integrity_check: "PASS" if new_md5 == original_md5 else "FAIL"}])
+            summary_df = pd.concat([summary_df, new_df], axis=0, ignore_index=True) #using concat as df.append is deprecated
+
+
+    print(summary_df)
+
+    # obtaining two lists of filenames that passed/failed verification, and moving to appropriate dir
+    for pfile in summary_df.query('file_integrity_check == "PASS"')['file_name'].to_list():
+        os.rename(f'{outdir}/{pfile}', f'{outdir}/{trans_int_outdir}/pass/{pfile}')
+
+    for ffile in summary_df.query('file_integrity_check == "FAIL"')['file_name'].to_list():
+        os.rename(f'{outdir}/{ffile}', f'{outdir}/{trans_int_outdir}/fail/{ffile}')
+
+    # creating md5 dir and moving all md5 files there
+    os.mkdir(f'{outdir}/{trans_int_outdir}/md5') # creating md5 dir
+    for file in [f for f in os.listdir(os.path.abspath(outdir)) if fnmatch.fnmatch(f, '*.md5*')]:
+        os.rename(f'{outdir}/{file}', f'{outdir}/{trans_int_outdir}/md5/{file}') # moving md5 files to md5 dir
+
+    # error check that count of files in pass + fail dir = total in summary_df
+    try:
+        len(os.listdir(pdir)) + len(os.listdir(fdir)) == len(summary_df)
+    except:
+        print("ERROR in transferring files to PASS/FAIL output directories", file=sys.stderr)
+
+    # creating output tsv file
+    summary_file_basename = f"file_integrity_summary_{dt_string}.tsv"
+    summary_filename = os.path.join(f'{outdir}/{trans_int_outdir}', summary_file_basename)
+    summary_df.to_csv(path_or_buf = summary_filename, sep = "\t",index = False)
+
+    # check for any corrupt data files
+    if (summary_df[integrity_check] == "PASS").all():
+        print(f"\nAll {len(summary_df)} data files in upload area: {outdir} have passed verification")
+    else:
+        print(f"\nERROR: some data files in upload area: {outdir} failed verification. Please see summary file {summary_file_basename}")
+
+
+def main():
+    transfer()
+
+if __name__ == '__main__':
+    main()
+
+
+
+# ### To test:
+# # gzip data files in pass dir
+# # todo: test if this script works as a re-run
+# # # TODO: test if this can run locally or not
+# #------
+#
