+++ modules/metadata_submission.nf
@@ -26,13 +26,13 @@ webin_account_mod = (params.webin_account.minus('su-'))
 
 if ( params.environment.toLowerCase() == 'test') {
 	"""
-	metadata_submission -f $spreadsheet -u $webin_account_mod -p $webin_password -a $action -o $PWD/$output -t
+	metadata_submission.py -f $spreadsheet -u $webin_account_mod -p $webin_password -a $action -o $PWD/$output -t
 	"""
  }
 
  else if (params.environment.toLowerCase() == 'prod' || params.environment.toLowerCase() == 'production') {
 	"""
-	metadata_submission -f $spreadsheet -u $webin_account_mod -p $webin_password -a $action -o $PWD/$output
+	metadata_submission.py -f $spreadsheet -u $webin_account_mod -p $webin_password -a $action -o $PWD/$output
 	"""
  }
