+++ nextflow.config
@@ -1,0 +1,16 @@
+profiles {
+    conda {
+        process.conda = "${projectDir}/package-list.yaml"
+}
+}
+conda.enabled = true
+params.action= "add"
+params.xml_output= "${projectDir}/output"
+params.webinCli_dir= "${projectDir}/webin-cli"
+params.sender_email= ""
+params.rec_email= ""
+params.transfer_output= "${projectDir}/data"
+
+
+
+
