--- pipeline/workflow/drag_and_drop_workflow/nextflow.config
@@ -1,10 +1,0 @@
-params.action= "add"
-params.xml_output= "${projectDir}/../../../output"
-params.webinCli_dir= "${projectDir}/../../../webin-cli"
-params.sender_email= ""
-params.rec_email= ""
-params.transfer_output= "${projectDir}/../../../data"
-
-
-
-
